[Feature] Auth data added to redux store
